+++ changelog.md
@@ -13,6 +13,8 @@ For example:
 
 ### Bug Fixes
 
+* Fixed API docs for creating and updating rows are missing for Multiple Select and Multiple Collaborator fields. [#1196](https://gitlab.com/bramw/baserow/-/issues/1196)
+
 ### Refactors
 
 ### Breaking API changes